Implemented QuickForm and QuickAction with some fix
