+++ frontend/src/features/pricing/pricing-group-form.schema.ts
@@ -1,15 +1,15 @@
-import { zPricingGroupControllerCreateData } from '@/integrations/api/client/zod.gen'
+import {
+  zPricingDto,
+  zPricingGroupControllerCreateData,
+} from '@/integrations/api/client/zod.gen'
 import z from 'zod'
 
 const zodPricingGroupCreate = zPricingGroupControllerCreateData.shape.body.shape
 const zodPricingGroup = zodPricingGroupCreate.group.shape
 
 export const pricingGroupFormSchema = z.object({
-  ...zodPricingGroupCreate,
-  group: {
-    name: zodPricingGroup.name.nonempty('Name should not be empty'),
-    amount: zodPricingGroup.amount.nonempty('Should have an amount'),
-  },
+  name: zodPricingGroup.name.nonempty('Name should not be empty'),
+  pricings: z.array(zPricingDto),
 })
 
 export type PricingGroupFormInput = z.input<typeof pricingGroupFormSchema>
