+++ frontend/src/pages/admin/pricing/index.tsx
@@ -1,10 +1,17 @@
+import GroupedIncludedFeesForm from '@/features/pricing/components/grouped-included-fees'
 import {
   pricingFormSchema,
   type PricingFormInput,
   type PricingFormOutput,
 } from '@/features/pricing/pricing-form.schema'
+import {
+  pricingGroupFormSchema,
+  type PricingGroupFormInput,
+  type PricingGroupFormOutput,
+} from '@/features/pricing/pricing-group-form.schema'
+import { useQuickAction } from '@/hooks/use-quick-action'
 import { useQuickForm } from '@/hooks/use-quick-form'
-import type { PricingDto, PricingType } from '@/integrations/api/client'
+import type { PricingType } from '@/integrations/api/client'
 import {
   pricingControllerCreateMutation,
   pricingControllerFindAllOptions,
@@ -16,18 +23,14 @@ import {
   pricingGroupControllerFindAllOptions,
   pricingGroupControllerFindAllQueryKey,
   pricingGroupControllerFindOneOptions,
-  pricingGroupControllerRemoveMutation,
   pricingGroupControllerUpdateMutation,
 } from '@/integrations/api/client/@tanstack/react-query.gen'
-import { getContext } from '@/integrations/tanstack-query/root-provider'
-import { useAppMutation } from '@/integrations/tanstack-query/useAppMutation'
+
 import {
   ActionIcon,
   Badge,
   Box,
   Button,
-  Card,
-  Combobox,
   Container,
   Divider,
   Drawer,
@@ -42,23 +45,19 @@ import {
   Text,
   TextInput,
   Title,
-  useCombobox,
 } from '@mantine/core'
-import { useForm } from '@mantine/form'
-import { useDebouncedValue, useInputState } from '@mantine/hooks'
 import {
   IconCheck,
   IconEdit,
-  IconMinus,
   IconPlus,
   IconSearch,
   IconTrash,
 } from '@tabler/icons-react'
-import { useQuery, useSuspenseQuery } from '@tanstack/react-query'
-import { getRouteApi, Router } from '@tanstack/react-router'
+import { useSuspenseQuery } from '@tanstack/react-query'
+import { getRouteApi } from '@tanstack/react-router'
 import Decimal from 'decimal.js'
 import { zod4Resolver } from 'mantine-form-zod-resolver'
-import { Suspense, useEffect, useState } from 'react'
+import { Suspense } from 'react'
 import z from 'zod'
 
 const route = getRouteApi('/(protected)/pricing/')
@@ -113,7 +112,7 @@ function GroupedFeesTab() {
           leftSection={<IconSearch size={18} stroke={1} />}
           w={rem(250)}
         />
-        <GroupedFeeForm />
+        <GroupedFeeFormDrawer />
       </Group>
       <GroupedFeesList />
     </Stack>
@@ -135,35 +134,16 @@ function GroupedFeesList() {
 
   const { pricingGroups, meta } = data
 
-  const { mutateAsync: deleteItem } = useAppMutation(
-    pricingGroupControllerRemoveMutation,
-    {
-      loading: {
-        title: 'Deleting pricing group',
-        message: 'Please wait while the pricing group is being deleted.',
-      },
-      success: {
-        title: 'Pricing Group Deleted',
-        message: 'The pricing group has been deleted.',
-      },
-      error: {
-        title: 'Failed',
-        message: 'Something went wrong while deleting the pricing group.',
-      },
-    },
-    {
-      onSuccess: () => {
-        const { queryClient } = getContext()
-
-        queryClient.invalidateQueries({
-          queryKey: pricingGroupControllerFindAllQueryKey(),
-        })
-      },
-    },
-  )
+  const { remove } = useQuickAction({
+    name: 'pricing',
+    removeMutationOptions: pricingControllerRemoveMutation({}),
+    queryKeyInvalidation: pricingControllerFindAllQueryKey({
+      // query: { page, search },
+    }),
+  })
 
   const handleDeletePricingGroup = async (id: string) => {
-    await deleteItem({
+    await remove.mutateAsync({
       path: { id },
       query: { directDelete: true },
     })
@@ -258,7 +238,7 @@ function GroupedFeesList() {
   )
 }
 
-function GroupedFeeForm() {
+function GroupedFeeFormDrawer() {
   const { createGroup, updateGroup } = route.useSearch()
   const navigate = route.useNavigate()
 
@@ -268,146 +248,6 @@ function GroupedFeeForm() {
         createGroup: state ? state : undefined,
       },
     })
-  }
-
-  const uuidSchema = z.uuidv4()
-
-  const { data: pricingGroup, isLoading } = useQuery({
-    ...pricingGroupControllerFindOneOptions({
-      path: { id: updateGroup || '' },
-    }),
-    enabled: uuidSchema.safeParse(updateGroup).success,
-  })
-
-  const form = useForm({
-    mode: 'uncontrolled',
-    initialValues: {
-      name: '',
-      includedFees: [] as PricingDto[],
-    },
-  })
-
-  useEffect(() => {
-    if (pricingGroup) {
-      form.setValues({
-        name: pricingGroup.name,
-        includedFees: pricingGroup.prices,
-      })
-    } else {
-      form.setValues({
-        name: '',
-        includedFees: [],
-      })
-    }
-  }, [pricingGroup])
-
-  const { mutateAsync: create, isPending: createPending } = useAppMutation(
-    pricingGroupControllerCreateMutation,
-    {
-      loading: {
-        title: 'Creating pricing group',
-        message: 'Please wait while the pricing group is being created.',
-      },
-      success: {
-        title: 'Pricing Group Created',
-        message: 'The pricing group has been created.',
-      },
-      error: {
-        title: 'Failed',
-        message: 'Something went wrong while creating the pricing group.',
-      },
-    },
-    {
-      onSuccess: () => {
-        const { queryClient } = getContext()
-
-        queryClient.invalidateQueries({
-          queryKey: pricingGroupControllerFindAllQueryKey(),
-        })
-      },
-    },
-  )
-
-  const handleCreatePricingGroup = async () => {
-    if (form.validate().hasErrors) return
-
-    const { name, includedFees } = form.getValues()
-
-    const amount = includedFees
-      .reduce(
-        (acc, val) => new Decimal(acc).add(new Decimal(val.amount)),
-        new Decimal(0),
-      )
-      .toString()
-
-    const pricings = includedFees.map((fee) => fee.id)
-
-    await create({
-      body: {
-        group: {
-          name,
-          amount,
-        },
-        pricings,
-      },
-    })
-
-    handleDrawerState(false)
-  }
-
-  const { mutateAsync: update, isPending: updatePending } = useAppMutation(
-    pricingGroupControllerUpdateMutation,
-    {
-      loading: {
-        title: 'Updating pricing group',
-        message: 'Please wait while the pricing group is being updated.',
-      },
-      success: {
-        title: 'Pricing Group Updated',
-        message: 'The pricing group has been updated.',
-      },
-      error: {
-        title: 'Failed',
-        message: 'Something went wrong while updating the pricing group.',
-      },
-    },
-    {
-      onSuccess: () => {
-        const { queryClient } = getContext()
-
-        queryClient.invalidateQueries({
-          queryKey: pricingGroupControllerFindAllQueryKey(),
-        })
-      },
-    },
-  )
-
-  const handleUpdatePricingGroup = async () => {
-    if (!updateGroup || form.validate().hasErrors) return
-
-    const { name, includedFees } = form.getValues()
-
-    const amount = includedFees
-      .reduce(
-        (acc, val) => new Decimal(acc).add(new Decimal(val.amount)),
-        new Decimal(0),
-      )
-      .toString()
-
-    const pricings = includedFees.map((fee) => fee.id)
-
-    await update({
-      path: { id: updateGroup },
-      body: {
-        group: {
-          name,
-          amount,
-        },
-        pricings,
-      },
-    })
-
-    handleDrawerState(false)
   }
 
   return (
@@ -435,210 +275,156 @@ function GroupedFeeForm() {
         // overlayProps={{ opacity: 0.2, blur: 2 }}
         padding="xl"
       >
-        <Stack gap="md">
-          <TextInput
-            label="Template Name"
-            placeholder="Name"
-            radius="md"
-            withAsterisk
-            disabled={isLoading || createPending || updatePending}
-            key={form.key('name')}
-            {...form.getInputProps('name')}
-          />
-          <GroupedIncludedFeesForm
-            disabled={isLoading || createPending || updatePending}
-            defaultFees={form.getValues().includedFees}
-            onAdd={(id) => {
-              form.setFieldValue('includedFees', [
-                ...form.getValues().includedFees,
-                id,
-              ])
-            }}
-            onRemove={(id) => {
-              form.setFieldValue(
-                'includedFees',
-                form.getValues().includedFees.filter((item) => item !== id),
-              )
-            }}
-          />
-          <Group justify="flex-end" mt="md">
-            <Button
-              variant="subtle"
-              disabled={isLoading || createPending || updatePending}
-              onClick={() => handleDrawerState(false)}
-            >
-              Cancel
-            </Button>
-            <Button
-              type="submit"
-              disabled={isLoading || createPending || updatePending}
-              onClick={() =>
-                createGroup
-                  ? handleCreatePricingGroup()
-                  : handleUpdatePricingGroup()
-              }
-            >
-              Save Changs
-            </Button>
-          </Group>
-        </Stack>
+        <GroupedFeeForm />
       </Drawer>
     </>
   )
 }
 
-interface GroupedIncludedFeesFormProps {
-  disabled: boolean
-  defaultFees?: PricingDto[]
-  onAdd: (pricing: PricingDto) => void
-  onRemove: (pricing: PricingDto) => void
-}
+function GroupedFeeForm() {
+  const { createGroup, updateGroup } = route.useSearch()
+  const navigate = route.useNavigate()
 
-function GroupedIncludedFeesForm({
-  disabled,
-  defaultFees,
-  onAdd,
-  onRemove,
-}: GroupedIncludedFeesFormProps) {
-  const [fees, setFees] = useState(defaultFees || [])
+  const handleDrawerState = (state: boolean) => {
+    navigate({
+      search: {
+        createGroup: state ? state : undefined,
+      },
+    })
+  }
 
-  useEffect(() => {
-    if (defaultFees) setFees(defaultFees)
-  }, [defaultFees])
+  const uuidSchema = z.uuidv4()
 
-  const combobox = useCombobox({
-    onDropdownClose: () => combobox.resetSelectedOption(),
-  })
-  const [value, setValue] = useInputState('')
-
-  const [debouncedSearch] = useDebouncedValue(value, 300)
-  const { data, isLoading } = useQuery({
-    ...pricingControllerFindAllOptions({
-      query: { search: debouncedSearch !== '' ? debouncedSearch : undefined },
+  const { create, update, form, isPending } = useQuickForm<
+    PricingGroupFormInput,
+    PricingGroupFormOutput
+  >()({
+    name: 'pricing group',
+    formOptions: {
+      initialValues: {
+        name: '',
+        pricings: [],
+      },
+      validate: zod4Resolver(pricingGroupFormSchema),
+    },
+    transformQueryData: (pricingGroup) => ({
+      name: pricingGroup.name,
+      pricings: pricingGroup.prices,
     }),
-    enabled: combobox.dropdownOpened,
+    queryOptions: {
+      ...pricingGroupControllerFindOneOptions({
+        path: { id: updateGroup || '' },
+      }),
+      enabled: uuidSchema.safeParse(updateGroup).success,
+    },
+    createMutationOptions: pricingGroupControllerCreateMutation({}),
+    updateMutationOptions: pricingGroupControllerUpdateMutation({
+      path: { id: updateGroup || '' },
+    }),
+    queryKeyInvalidation: pricingGroupControllerFindAllQueryKey({
+      // query: { page, search },
+    }),
   })
 
-  const shouldFilterOptions =
-    data && !data.pricings.some((item) => item.name === value)
-  const filteredOptions = (
-    shouldFilterOptions
-      ? data.pricings.filter((item) =>
-          item.name.toLowerCase().includes(value.toLowerCase().trim()),
-        )
-      : data?.pricings
-  )?.filter((item) => !fees.some((fee) => fee.id === item.id))
+  const handleCreatePricingGroup = async () => {
+    if (form.validate().hasErrors) return
 
-  useEffect(() => {
-    combobox.selectFirstOption()
-  }, [value])
+    const { name, pricings: includedFees } = form.getValues()
+
+    const amount = includedFees
+      .reduce(
+        (acc, val) => new Decimal(acc).add(new Decimal(val.amount)),
+        new Decimal(0),
+      )
+      .toString()
+
+    const pricings = includedFees.map((fee) => fee.id)
+
+    await create.mutateAsync({
+      body: {
+        group: {
+          name,
+          amount,
+        },
+        pricings,
+      },
+    })
+
+    handleDrawerState(false)
+  }
+
+  const handleUpdatePricingGroup = async () => {
+    if (!updateGroup || form.validate().hasErrors) return
+
+    const { name, pricings: includedFees } = form.getValues()
+
+    const amount = includedFees
+      .reduce(
+        (acc, val) => new Decimal(acc).add(new Decimal(val.amount)),
+        new Decimal(0),
+      )
+      .toString()
+
+    const pricings = includedFees.map((fee) => fee.id)
+
+    await update.mutateAsync({
+      path: { id: updateGroup },
+      body: {
+        group: {
+          name,
+          amount,
+        },
+        pricings,
+      },
+    })
+
+    handleDrawerState(false)
+  }
 
   return (
-    <Stack>
-      <Combobox
-        onOptionSubmit={(optionValue) => {
-          if (data) {
-            const found = data.pricings.find(
-              (price) => price.id === optionValue,
-            )
-
-            if (found) {
-              setFees((prev) => [...prev, found])
-              onAdd(found)
-            }
-          }
-          setValue('')
-          combobox.closeDropdown()
+    <Stack gap="md">
+      <TextInput
+        label="Template Name"
+        placeholder="Name"
+        radius="md"
+        withAsterisk
+        disabled={isPending}
+        key={form.key('name')}
+        {...form.getInputProps('name')}
+      />
+      <GroupedIncludedFeesForm
+        disabled={isPending}
+        defaultFees={form.getValues().pricings}
+        onAdd={(id) => {
+          form.setFieldValue('includedFees', [...form.getValues().pricings, id])
         }}
-        store={combobox}
-      >
-        <Combobox.Target>
-          <TextInput
-            label="Included Fees"
-            placeholder="Search"
-            leftSection={<IconSearch size={18} stroke={1} />}
-            value={value}
-            onChange={(event) => {
-              setValue(event)
-              combobox.openDropdown()
-              combobox.updateSelectedOptionIndex()
-            }}
-            onClick={() => combobox.openDropdown()}
-            onFocus={() => combobox.openDropdown()}
-            onBlur={() => combobox.closeDropdown()}
-            flex={1}
-            disabled={disabled}
-          />
-        </Combobox.Target>
-
-        <Combobox.Dropdown>
-          <Combobox.Options>
-            {isLoading ? (
-              <Combobox.Empty>Fetching...</Combobox.Empty>
-            ) : filteredOptions?.length === 0 ? (
-              <Combobox.Empty>Nothing found</Combobox.Empty>
-            ) : (
-              filteredOptions?.map((pricing) => (
-                <Combobox.Option value={pricing.id} key={pricing.id}>
-                  <Group justify="space-between">
-                    <Text>{pricing.name}</Text>
-                    <Text>₱{pricing.amount}</Text>
-                  </Group>
-                </Combobox.Option>
-              ))
-            )}
-          </Combobox.Options>
-        </Combobox.Dropdown>
-      </Combobox>
-      <Stack gap="xs">
-        {fees.map((fee) => (
-          <Group key={fee.id}>
-            <Card radius="md" py={8} withBorder flex={1}>
-              <Group justify="space-between">
-                <Stack gap={2}>
-                  <Text size="sm">{fee.name}</Text>
-                  <Text size="xs" c="dimmed" tt="capitalize">
-                    {fee.type}
-                  </Text>
-                </Stack>
-                <Text size="sm">₱{fee.amount}</Text>
-              </Group>
-            </Card>
-            <ActionIcon
-              size="sm"
-              variant="outline"
-              radius="lg"
-              onClick={() => {
-                onRemove(fee)
-                setFees((prev) => prev.filter((item) => item.id !== fee.id))
-              }}
-              disabled={disabled}
-            >
-              <IconMinus />
-            </ActionIcon>
-          </Group>
-        ))}
-        <Divider />
-        <Group mr={36}>
-          <Card radius="md" py={8} flex={1}>
-            <Group justify="space-between">
-              <Text size="sm" fw={500}>
-                Total
-              </Text>
-
-              <Text size="sm" fw={500}>
-                ₱
-                {fees
-                  .reduce(
-                    (acc, val) => new Decimal(acc).add(new Decimal(val.amount)),
-                    new Decimal(0),
-                  )
-                  .toString()}
-              </Text>
-            </Group>
-          </Card>
-        </Group>
-      </Stack>
+        onRemove={(id) => {
+          form.setFieldValue(
+            'includedFees',
+            form.getValues().pricings.filter((item) => item !== id),
+          )
+        }}
+      />
+      <Group justify="flex-end" mt="md">
+        <Button
+          variant="subtle"
+          disabled={isPending}
+          onClick={() => handleDrawerState(false)}
+        >
+          Cancel
+        </Button>
+        <Button
+          type="submit"
+          disabled={isPending}
+          onClick={() =>
+            createGroup
+              ? handleCreatePricingGroup()
+              : handleUpdatePricingGroup()
+          }
+        >
+          Save Changs
+        </Button>
+      </Group>
     </Stack>
   )
 }
@@ -653,7 +439,7 @@ function IndividualFeesTab() {
           leftSection={<IconSearch size={18} stroke={1} />}
           w={rem(250)}
         />
-        <IndividualFeeForm />
+        <IndividualFeeFormDrawer />
       </Group>
       <IndividualFeesList />
     </Stack>
@@ -675,35 +461,16 @@ function IndividualFeesList() {
 
   const { pricings, meta } = data
 
-  const { mutateAsync: deleteItem } = useAppMutation(
-    pricingControllerRemoveMutation,
-    {
-      loading: {
-        title: 'Deleting pricing',
-        message: 'Please wait while the pricing is being deleted.',
-      },
-      success: {
-        title: 'Pricing Deleted',
-        message: 'The pricing has been deleted.',
-      },
-      error: {
-        title: 'Failed',
-        message: 'Something went wrong while deleting the pricing.',
-      },
-    },
-    {
-      onSuccess: () => {
-        const { queryClient } = getContext()
-
-        queryClient.invalidateQueries({
-          queryKey: pricingControllerFindAllQueryKey(),
-        })
-      },
-    },
-  )
+  const { remove } = useQuickAction({
+    name: 'pricing',
+    removeMutationOptions: pricingControllerRemoveMutation({}),
+    queryKeyInvalidation: pricingControllerFindAllQueryKey({
+      // query: { page, search },
+    }),
+  })
 
   const handleDeletePricing = async (id: string) => {
-    await deleteItem({
+    await remove.mutateAsync({
       path: { id },
       query: { directDelete: true },
     })
@@ -769,6 +536,47 @@ function IndividualFeesList() {
   )
 }
 
+function IndividualFeeFormDrawer() {
+  const { createFee, updateFee } = route.useSearch()
+  const navigate = route.useNavigate()
+
+  const handleDrawerState = (state: boolean) => {
+    navigate({
+      search: {
+        createFee: state ? state : undefined,
+      },
+    })
+  }
+  return (
+    <>
+      <Button
+        variant="filled"
+        radius={'md'}
+        leftSection={<IconPlus size={20} />}
+        lts={rem(0.25)}
+        onClick={() => handleDrawerState(true)}
+      >
+        Create
+      </Button>
+
+      <Drawer
+        opened={createFee === true || updateFee !== undefined}
+        onClose={() => handleDrawerState(false)}
+        title={
+          <Text size="xl" fw={600}>
+            {createFee ? 'Create' : 'Update'} Individual Fee
+          </Text>
+        }
+        position="right"
+        size="md"
+        padding="xl"
+      >
+        <IndividualFeeForm />
+      </Drawer>
+    </>
+  )
+}
+
 function IndividualFeeForm() {
   const { createFee, updateFee } = route.useSearch()
   const navigate = route.useNavigate()
@@ -785,7 +593,7 @@ function IndividualFeeForm() {
     PricingFormInput,
     PricingFormOutput
   >()({
-    name: 'course',
+    name: 'pricing',
     formOptions: {
       initialValues: {
         name: '',
@@ -813,8 +621,6 @@ function IndividualFeeForm() {
       // query: { page, search },
     }),
   })
-
-  console.log(form.getValues())
 
   const handleCreatePricing = async () => {
     if (form.validate().hasErrors) return
@@ -861,78 +667,53 @@ function IndividualFeeForm() {
   ]
 
   return (
-    <>
-      <Button
-        variant="filled"
-        radius={'md'}
-        leftSection={<IconPlus size={20} />}
-        lts={rem(0.25)}
-        onClick={() => handleDrawerState(true)}
-      >
-        Create
-      </Button>
-
-      <Drawer
-        opened={createFee === true || updateFee !== undefined}
-        onClose={() => handleDrawerState(false)}
-        title={
-          <Text size="xl" fw={600}>
-            {createFee ? 'Create' : 'Update'} Individual Fee
-          </Text>
-        }
-        position="right"
-        size="md"
-        padding="xl"
-      >
-        <Stack gap="md">
-          <TextInput
-            label="Template Name"
-            placeholder="Name"
-            radius="md"
-            withAsterisk
-            disabled={isPending}
-            key={form.key('name')}
-            {...form.getInputProps('name')}
-          />
-          <Select
-            label="Fee Type"
-            placeholder="Select fee type"
-            data={pricingTypes}
-            withAsterisk
-            disabled={isPending}
-            key={form.key('type')}
-            {...form.getInputProps('type')}
-          />
-          <TextInput
-            label="Fee Amount (₱)"
-            placeholder="0.00"
-            type="number"
-            radius="md"
-            required
-            min={0}
-            disabled={isPending}
-            key={form.key('amount')}
-            {...form.getInputProps('amount')}
-          />
-          <Group justify="flex-end" mt="md">
-            <Button
-              variant="subtle"
-              onClick={() => handleDrawerState(false)}
-              disabled={isPending}
-            >
-              Cancel
-            </Button>
-            <Button
-              onClick={() =>
-                createFee ? handleCreatePricing() : handleUpdatePricing()
-              }
-              disabled={isPending}
-            >
-              Save Changes
-            </Button>
-          </Group>
-        </Stack>
-      </Drawer>
-    </>
+    <Stack gap="md">
+      <TextInput
+        label="Template Name"
+        placeholder="Name"
+        radius="md"
+        withAsterisk
+        disabled={isPending}
+        key={form.key('name')}
+        {...form.getInputProps('name')}
+      />
+      <Select
+        label="Fee Type"
+        placeholder="Select fee type"
+        data={pricingTypes}
+        withAsterisk
+        disabled={isPending}
+        key={form.key('type')}
+        {...form.getInputProps('type')}
+      />
+      <TextInput
+        label="Fee Amount (₱)"
+        placeholder="0.00"
+        type="number"
+        radius="md"
+        required
+        min={0}
+        disabled={isPending}
+        key={form.key('amount')}
+        {...form.getInputProps('amount')}
+      />
+      <Group justify="flex-end" mt="md">
+        <Button
+          variant="subtle"
+          onClick={() => handleDrawerState(false)}
+          disabled={isPending}
+        >
+          Cancel
+        </Button>
+        <Button
+          onClick={() =>
+            createFee ? handleCreatePricing() : handleUpdatePricing()
+          }
+          disabled={isPending}
+        >
+          Save Changes
+        </Button>
+      </Group>
+    </Stack>
   )
 }
